requirements.txt
numpy>=1.17.4
pycryptodome>=3.9.6
Pillow>=7.0.0
argon2
gnupg
